friendly blue cube come back & cell updater was updated
cell updateGeometry added building in field.

blu cube was back and small change.

wall price has been defined how integer.

+++ src/main/java/io/github/sevjet/essensedefence/Wall.java
@@ -2,13 +2,13 @@ package io.github.sevjet.essensedefence;
 
 public class Wall extends Building {
 
-    private double price;
+    private int price;
 
     public Wall() {
-        this(0.0D);
+        this(0);
     }
 
-    public Wall(double price) {
+    public Wall(int price) {
         setHeight(1);
         setWidth(1);
 
@@ -20,11 +20,11 @@ public class Wall extends Building {
 
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
